[QCF-69] 게시글, 장비, 좋아요 모아보기 데이터 NavigationTab으로 이동
