+++ desk/src/components/units/user/User.types.ts
@@ -6,14 +6,10 @@ export type UserProps = {
 
 export type UserUIProps = {
   userData: TFetchUser
-  showUserPosts: boolean
-  showUserProductPosts: boolean
-  showLikedPosts: boolean
   isLoggedIn: boolean
   isMyPage: boolean
   isLiked: boolean
   toggleIsLiked: () => void
-  onClickTab: (id: number) => void
   onClickMoveToAccountEdit: () => void
   onClickFollowingButton: () => void
 }
